LayoutTests/imported/w3c/web-platform-tests/FileAPI/url/url-format.any.worker-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: File

FAIL url-format ReferenceError: Can't find variable: File
